moved repititive sections to a function
