+++ steward/src/commands/schedule_corks/aave_cellar/shutdown.rs
@@ -1,15 +1,6 @@
-use crate::{
-    application::APP,
-    cellars::{self},
-    config,
-    prelude::*,
-    somm_send,
-    utils::{CHAIN_PREFIX, MESSAGE_TIMEOUT},
-};
-use abscissa_core::{clap::Parser, Application, Command, Runnable};
-use deep_space::{Coin, Contact};
+use crate::{application::APP, prelude::*, utils::SubmitCork};
+use abscissa_core::{clap::Parser, Command, Runnable};
 use ethers::abi::AbiEncode;
-use somm_proto::cork::Cork;
 use steward_abi::aave_v2_stablecoin::*;
 
 /// Shutdown subcommand
@@ -37,8 +28,6 @@ pub struct ShutdownCmd {
 
 impl Runnable for ShutdownCmd {
     fn run(&self) {
-        let config = APP.config();
-
         abscissa_tokio::run_with_actix(&APP, async {
             let call = SetShutdownCall {
                 shutdown: self.shut_down,
@@ -46,35 +35,13 @@ impl Runnable for ShutdownCmd {
             };
             let encoded_call = AaveV2StablecoinCellarCalls::SetShutdown(call).encode();
 
-            cellars::validate_cellar_id(self.contract.as_str()).unwrap_or_else(|err| {
-                status_err!("Can't validate contract address format: {}", err);
-                std::process::exit(1);
-            });
-
-            let cork = Cork {
-                encoded_contract_call: encoded_call,
-                target_contract_address: self.contract.clone(),
+            let submit = SubmitCork {
+                contract: self.contract.clone(),
+                height: self.height,
+                encoded_call,
             };
 
-            debug!("establishing grpc connection");
-            let contact = Contact::new(&config.cosmos.grpc, MESSAGE_TIMEOUT, CHAIN_PREFIX).unwrap();
-
-            debug!("getting cosmos fee");
-            let cosmos_gas_price = config.cosmos.gas_price.as_tuple();
-            let fee = Coin {
-                amount: (cosmos_gas_price.0 as u64).into(),
-                denom: cosmos_gas_price.1,
-            };
-            somm_send::schedule_cork(
-                &contact,
-                cork,
-                config::DELEGATE_ADDRESS.to_string(),
-                &config::DELEGATE_KEY,
-                fee,
-                self.height,
-            )
-            .await
-            .unwrap_or_else(|err| {
+            submit.submit_cork().await.unwrap_or_else(|err| {
                 status_err!("executor exited with error: {}", err);
                 std::process::exit(1);
             })
