+++ steward/src/utils.rs
@@ -1,17 +1,26 @@
 use crate::{
+    application::APP,
+    cellars::{self},
+    config,
     error::{Error, ErrorKind},
-    prelude::APP,
+    prelude::*,
+    somm_send,
 };
 use abscissa_core::Application;
 use deep_space::error::CosmosGrpcError;
+use deep_space::{Coin, Contact};
 use ethers::prelude::{types::Address as EthAddress, *};
 use gravity_bridge::{
-    gravity_proto::gravity::{
-        query_client::QueryClient, DelegateKeysByOrchestratorRequest,
-        DelegateKeysByOrchestratorResponse,
+    gravity_proto::{
+        cosmos_sdk_proto::cosmos::base::abci::v1beta1::TxResponse,
+        gravity::{
+            query_client::QueryClient, DelegateKeysByOrchestratorRequest,
+            DelegateKeysByOrchestratorResponse,
+        },
     },
     gravity_utils::ethereum::downcast_to_u64,
 };
+use somm_proto::cork::Cork;
 use std::{convert::TryFrom, time::Duration};
 use tonic::transport::Channel;
 
@@ -89,4 +98,50 @@ pub async fn get_eth_provider() -> Result<Provider<Http>, Error> {
 
 pub fn sp_call_error(message: String) -> Error {
     ErrorKind::SPCallError.context(message).into()
+}
+
+pub struct SubmitCork {
+    pub contract: String,
+    pub height: u64,
+    pub encoded_call: Vec<u8>,
+}
+
+impl SubmitCork {
+    pub async fn submit_cork(&self) -> Result<TxResponse, CosmosGrpcError> {
+        let config = APP.config();
+        // Validate cellar id
+        cellars::validate_cellar_id(self.contract.as_str()).unwrap_or_else(|err| {
+            status_err!("Can't validate contract address format: {}", err);
+            std::process::exit(1);
+        });
+
+        let cork = Cork {
+            encoded_contract_call: self.encoded_call.clone(),
+            target_contract_address: self.contract.clone(),
+        };
+
+        // Establish grpc connections
+        debug!("establishing grpc connection");
+        let contact = Contact::new(&config.cosmos.grpc, MESSAGE_TIMEOUT, CHAIN_PREFIX).unwrap();
+
+        // Get cosmos fees
+        debug!("getting cosmos fee");
+        let cosmos_gas_price = config.cosmos.gas_price.as_tuple();
+
+        let fee = Coin {
+            amount: (cosmos_gas_price.0 as u64).into(),
+            denom: cosmos_gas_price.1,
+        };
+
+        // send scheduled cork
+        somm_send::schedule_cork(
+            &contact,
+            cork,
+            config::DELEGATE_ADDRESS.to_string(),
+            &config::DELEGATE_KEY,
+            fee,
+            self.height,
+        )
+        .await
+    }
 }
